backend/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
celery>=5.2.0
redis>=4.0.0
httpx>=0.24.0
prometheus-client>=0.12.0
websockets>=10.0
python-dotenv>=0.19.0
bcrypt==3.2.2
email-validator>=2.0.0